Add files via upload
Cardiovascular diseases are the leading cause of death globally. To identify the causes and to develop a system to predict heart attack in an effective manner is necessary. The presented data has all the information about all the relevant factors that might have an impact on heart health. The data needs to be explained in detail for any further analysis.

1. Preliminary analysis:

Perform preliminary data inspection and report the findings as to the structure of the data, missing values, duplicates, etc.
Based on the findings from the previous question remove duplicates (if any) , treat missing values using an appropriate strategy.
2. Prepare an informative report about the data explaining the distribution of the disease and the related factors. You could use the below approach to achieve the objective

Get a preliminary statistical summary of the data. Explore the measures of central tendencies and the spread of the data overall.
Identify the data variables which might be categorical in nature. Describe and explore these variables using appropriate tools e.g. count plot
Study the occurrence of CVD across Age.
Study the composition of overall patients w.r.t. Gender.
Can we detect a heart attack based on anomalies in the Resting Blood Pressure of the patient?
Describe the relationship between Cholesterol levels and our target variable.
What can be concluded about the relationship between peak exercising and the occurrence of a heart attack.
Is thalassemia a major cause of CVD?
How are the other factors determining the occurrence of CVD?
Use a pair plot to understand the relationship between all the given variables.
3. Build a baseline model to predict using a Logistic Regression and explore the results.

 

Project Task: Week 1

Importing, Understanding, and Inspecting Data :

Perform preliminary data inspection and report the findings as the structure of the data, missing values, duplicates, etc.

Based on the findings from the previous question, remove duplicates (if any) and treat missing values using an appropriate strategy.

Get a preliminary statistical summary of the data. Explore the measures of central tendencies and the spread of the data overall.

Performing EDA:

Identify the data variables which might be categorical in nature. Describe and explore these variables using appropriate tools. For example: count plot.

Study the occurrence of CVD across different ages.

Can we detect heart attack based on anomalies in resting blood pressure of the patient?

Study the composition of overall patients w.r.t . gender.


 

Project Task: Week 2

Performing EDA and Modeling:

Describe the relationship between cholesterol levels and our target variable.

What can be concluded about the relationship between peak exercising and occurrence of heart attack?

Is thalassemia a major cause of CVD? How are the other factors determining the occurrence of CVD?

Use a pair plot to understand the relationship between all the given variables.

Perform logistic regression, predict the outcome for test data, and validate the results by using the confusion matrix.

Dashboarding:

Visualize the variables using Tableau to create an understanding for attributes of a Diseased vs. a Healthy person.

Demonstrate  the variables associated with each other and factors to build a dashboard